+++ idea/testData/quickfix/createFromUsage/createClass/typeReference/noObjectForTypeConstraint.kt
@@ -2,6 +2,7 @@
 // ACTION: Create class 'NotExistent'
 // ACTION: Create interface 'NotExistent'
 // ACTION: Create type alias 'NotExistent'
+// ACTION: Create type parameter 'NotExistent' in class 'TPB'
 // ACTION: Create test
 // ERROR: Unresolved reference: NotExistent
 class TPB<X> where X : <caret>NotExistent