+++ idea/testData/quickfix/createFromUsage/createTypeParameter/classNoExplication.kt
@@ -1,0 +1,13 @@
+// "Create type parameter 'X' in class 'Foo'" "true"
+open class Foo(x: <caret>X)
+
+class Bar : Foo(1)
+
+fun test() {
+    Foo(1)
+    Foo("2")
+
+    object : Foo("2") {
+
+    }
+}